TW-249: Find deadcode. -- References 'lib' -> 'app'
