--- src/app/hooks/useGasToken.ts
@@ -1,21 +1,0 @@
-import { useNetwork } from 'lib/temple/front';
-import { FILM_METADATA, TEZOS_METADATA } from 'lib/temple/metadata';
-
-export const useGasToken = () => {
-  const network = useNetwork();
-
-  return network.type === 'dcp'
-    ? {
-        logo: 'misc/token-logos/film.png',
-        symbol: 'ф',
-        assetName: 'FILM',
-        metadata: FILM_METADATA,
-        isDcpNetwork: true
-      }
-    : {
-        logo: 'misc/token-logos/tez.svg',
-        symbol: 'ꜩ',
-        assetName: 'tez',
-        metadata: TEZOS_METADATA
-      };
-};
