+++ src/app/pages/Explore/BakingHistoryItem.tsx
@@ -19,11 +19,10 @@ import { ReactComponent as TimeIcon } from 'app/icons/time.svg';
 import HashChip from 'app/templates/HashChip';
 import { getPluralKey, toLocalFormat } from 'lib/i18n/numbers';
 import { T } from 'lib/i18n/react';
-import { getRewardsStats, useKnownBaker, useExplorerBaseUrls } from 'lib/temple/front';
+import { getRewardsStats, useKnownBaker, useExplorerBaseUrls, useGasToken } from 'lib/temple/front';
 import { mutezToTz } from 'lib/temple/helpers';
 import { TzktRewardsEntry } from 'lib/tzkt';
 
-import { useGasToken } from '../../hooks/useGasToken';
 import styles from './BakingHistoryItem.module.css';
 
 type BakingHistoryItemProps = {
