+++ src/app/templates/AccountBanner.tsx
@@ -8,9 +8,8 @@ import Money from 'app/atoms/Money';
 import Name from 'app/atoms/Name';
 import Balance from 'app/templates/Balance';
 import { t } from 'lib/i18n/react';
+import { useGasToken } from 'lib/temple/front';
 import { TempleAccount } from 'lib/temple/types';
-
-import { useGasToken } from '../hooks/useGasToken';
 
 type AccountBannerProps = HTMLAttributes<HTMLDivElement> & {
   account: TempleAccount;
