+++ src/app/templates/AdditionalFeeInput.tsx
@@ -26,8 +26,8 @@ import CustomSelect, { OptionRenderProps } from 'app/templates/CustomSelect';
 import { AnalyticsEventCategory, useAnalytics } from 'lib/analytics';
 import { toLocalFixed } from 'lib/i18n/numbers';
 import { TID, T, t } from 'lib/i18n/react';
+import { useGasToken } from 'lib/temple/front';
 
-import { useGasToken } from '../hooks/useGasToken';
 import { AdditionalFeeInputSelectors } from './AdditionalFeeInput.selectors';
 
 type AssetFieldProps = typeof AssetField extends ForwardRefExoticComponent<infer T> ? T : never;
