+++ src/lib/env.ts
@@ -1,0 +1,1 @@
+export const IS_DEV_ENV = process.env.NODE_ENV === 'development';
