+++ src/lib/temple/front/ready.ts
@@ -5,7 +5,7 @@ import { TezosToolkit } from '@taquito/taquito';
 import { Tzip16Module } from '@taquito/tzip16';
 import constate from 'constate';
 
-import { IS_DEV_ENV } from 'app/env';
+import { IS_DEV_ENV } from 'lib/env';
 import { useRetryableSWR } from 'lib/swr';
 import { loadChainId, michelEncoder, loadFastRpcClient } from 'lib/temple/helpers';
 import { ReadyTempleState, TempleAccountType, TempleStatus, TempleState } from 'lib/temple/types';
